requirements.txt
python-telegram-bot
opencv-python==4.8.0.74
numpy==1.26.4
pyzbar
pillow
libzbar0